Cleaning up some type files
