+++ server/Controllers/StreamController.ts
@@ -30,35 +30,7 @@ export class StreamController implements IController {
 
     const vPath = path.join(mediaPath, video.filename);
 
-    const stat = fs.statSync(vPath);
-    const fileSize = stat.size;
-    const range = req.headers.range;
-
-    if (range) {
-      const parts = range.replace(/bytes=/, '').split('-');
-      const start = parseInt(parts[0], 10);
-      if (start === 0) await this._mediaManager.incrementViewCount(videoId);
-      const end = parts[1] ? parseInt(parts[1], 10) : fileSize - 1;
-
-      const chunksize = end - start + 1;
-      const file = fs.createReadStream(vPath, { start, end });
-      const head = {
-        'Content-Range': `bytes ${start}-${end}/${fileSize}`,
-        'Accept-Ranges': 'bytes',
-        'Content-Length': chunksize,
-        'Content-Type': 'video/mp4'
-      };
-
-      res.writeHead(206, head);
-      file.pipe(res);
-    } else {
-      const head = {
-        'Content-Length': fileSize,
-        'Content-Type': 'video/mp4'
-      };
-      res.writeHead(200, head);
-      fs.createReadStream(vPath).pipe(res);
-    }
+    res.sendFile(vPath);
   };
 
   getCaptions: RequestHandler = async (req, res) => {
