+++ server/models/MediaType.ts
@@ -1,0 +1,7 @@
+export enum MediaType {
+  Video = 'video',
+  Audio = 'audio',
+  Image = 'image',
+  Book = 'book',
+  Binary = 'binary'
+}
