+++ server/models/ServerLog.ts
@@ -1,0 +1,9 @@
+import { ServerLogSeverity } from './ServerLogSeverity';
+
+export type ServerLog = {
+  severity: ServerLogSeverity;
+  source?: string;
+  createdAt: Date;
+  message: string;
+  additional?: any;
+};
