+++ server/models/ServerLogSeverity.ts
@@ -1,0 +1,5 @@
+export enum ServerLogSeverity {
+  info,
+  warn,
+  err
+}
