--- server/models/SetMediaOrderPayload.ts
@@ -1,5 +1,0 @@
-export interface SetMediaOrderPayload {
-  libraryId: string;
-  mediaId: string;
-  direction: 'up' | 'down';
-}
